exploit: port to release kernel and fix bug in ptmx_fops repairment
This patch ports the exploit to the original 4.6 kernel
without any additional debug prints patched into it.

Furthermore, while porting the exploit, I have noticed that
I have made a small mistake when repairing the ptmx_fops
structure - the written memory layout of ptmx_fops was a
bit off. So I fixed that along the way.

Signed-off-by: Milo-D <milod2048@gmail.com>